Removed global, now running the blocks-static and app.
Removed the global viewReady method and added it to the browser object.

Removed the logic to spawn a blocks-static like server and the logic to
spawn a blocks-server.

Now instead we are spawning a new node instance with the blocks static or
the app.js.

For some reason tests are failing sometime.
Sometimes something get's wrong server side, sometimes we are running
into timeouts and sometimes there are client side errors.

I'll have to look into that later.